requirements.txt
annotated-types==0.5.0
anyio==4.0.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
flask==2.3.3
Flask-Cors==4.0.0
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.13.0
grpcio==1.59.0
grpcio-tools==1.59.0
gspread==5.12.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
llmsherpa==0.1.3
MarkupSafe==2.1.3
numpy==1.24.4
oauthlib==3.2.2
packaging==23.1
portalocker==2.8.2
proto-plus==1.23.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic-core==2.10.1
pyparsing==3.1.1
pypdf2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
qdrant-client==1.5.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
shapely==2.0.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
StrEnum==0.4.15
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.16
uuid==1.30
werkzeug==2.3.7
zipp==3.17.0
